azure-dental-onboarding-template/requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.21.0
